Fixed a typo in src/seedu/addressbook/AddressBook.java (AddessBook -> AddressBook)
Changed the internal representation of a person to use a Map instead of a String.
Removed the variable PERSON_DATA_COUNT.
Added a new private enum PersonProperty to function as the keys for the person Map representation.

Added a new test to test/input.txt
Added a new test outcome to test/expected.txt

Formatted tests/runtests.bat for ease of reading.